requirements.txt
pytesseract==0.3.10
pillow==9.5.0
opencv-python==4.7.0.72
easyocr==1.6.2
dask[complete]==2023.4.0
pandas==2.0.3
scikit-learn==1.2.2
pyzbar==0.1.9
flask==2.3.3
flask_cors==3.0.10
torch==2.0.1
langdetect==1.0.9
argostranslate==1.8.0
